slide-pc/pyproject.toml
[project]
name = "slide"
version = "0.1.0"
description = "SLIDE - Serial Line Inter-Device (file) Exchange"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
]

[project.scripts]
slide-send = "slide.send:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
